Renames handover_api->d4s2_api
