+++ d4s2_api/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from handover_api.models import *
+from d4s2_api.models import *
 from simple_history.admin import SimpleHistoryAdmin
 
 admin.site.register(DukeDSUser)
