--- handover_api/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'handover_api.apps.HandoverApiConfig'
